+++ apps/playground/CHANGELOG.md
@@ -1,5 +1,13 @@
 # @postrun/playground
 
+## 0.0.4
+
+### Patch Changes
+
+- Updated dependencies
+  - @postrun/react@2.6.0
+  - @postrun/js@2.6.0
+
 ## 0.0.3
 
 ### Patch Changes